# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResult {
  success: boolean;
  /** @nullable */
  error?: string | null;
}

export interface Guild {
  id: string;
  name: string;
  memberCount: number;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Channel {
  id: string;
  name: string;
  type: string;
  /** @nullable */
  topic?: string | null;
  position?: number;
}

export interface Role {
  id: string;
  name: string;
  color: string;
  position: number;
}

export interface GuildStats {
  totalMembers: number;
  totalChannels: number;
  totalRoles: number;
  onlineMembers: number;
}

export interface GuildSettings {
  /** @nullable */
  id?: number | null;
  guildId: string;
  /** @nullable */
  welcomeEnabled?: boolean | null;
  /** @nullable */
  welcomeChannelId?: string | null;
  /** @nullable */
  welcomeTitle?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  /** @nullable */
  welcomeColor?: string | null;
  /** @nullable */
  welcomeShowAvatar?: boolean | null;
  /** @nullable */
  autoRoleEnabled?: boolean | null;
  /** @nullable */
  autoRoleId?: string | null;
  /** @nullable */
  ticketEnabled?: boolean | null;
  /** @nullable */
  ticketChannelId?: string | null;
  /** @nullable */
  ticketCategory?: string | null;
  /** @nullable */
  ticketMessage?: string | null;
  /** @nullable */
  loggingEnabled?: boolean | null;
  /** @nullable */
  loggingChannelId?: string | null;
  /** @nullable */
  moderationEnabled?: boolean | null;
  /** @nullable */
  moderationChannels?: string | null;
  /** @nullable */
  antiSpamEnabled?: boolean | null;
  /** @nullable */
  antiSpamThreshold?: string | null;
  /** @nullable */
  xpEnabled?: boolean | null;
  /** @nullable */
  xpPerMessage?: string | null;
  /** @nullable */
  bannedWords?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface GuildSettingsInput {
  welcomeEnabled?: boolean;
  /** @nullable */
  welcomeChannelId?: string | null;
  welcomeTitle?: string;
  welcomeMessage?: string;
  welcomeColor?: string;
  welcomeShowAvatar?: boolean;
  autoRoleEnabled?: boolean;
  /** @nullable */
  autoRoleId?: string | null;
  ticketEnabled?: boolean;
  /** @nullable */
  ticketChannelId?: string | null;
  ticketCategory?: string;
  ticketMessage?: string;
  loggingEnabled?: boolean;
  /** @nullable */
  loggingChannelId?: string | null;
  moderationEnabled?: boolean;
  moderationChannels?: string;
  antiSpamEnabled?: boolean;
  antiSpamThreshold?: string;
  xpEnabled?: boolean;
  xpPerMessage?: string;
  bannedWords?: string;
}

export interface MessageInput {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface MessageResult {
  id: string;
  success: boolean;
}

export interface TicketPanelInput {
  guildId: string;
}

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface Ticket {
  id: number;
  channelId: string;
  userId: string;
  username: string;
  guildId: string;
  /** @nullable */
  category?: string | null;
  status: TicketStatus;
  createdAt: string;
  /** @nullable */
  closedAt?: string | null;
}

export interface TicketMessage {
  id: number;
  ticketId: number;
  userId: string;
  username: string;
  content: string;
  createdAt: string;
}

export interface Warning {
  id: number;
  userId: string;
  username: string;
  guildId: string;
  count: number;
  reason: string;
  createdAt: string;
}

export interface XpEntry {
  id: number;
  userId: string;
  username: string;
  guildId: string;
  xp: number;
  level: number;
  updatedAt?: string;
}

export interface Giveaway {
  id: number;
  /** @nullable */
  messageId?: string | null;
  channelId: string;
  guildId: string;
  prize: string;
  winnersCount: number;
  hostId: string;
  hostUsername: string;
  participants?: string[];
  winners?: string[];
  active: boolean;
  endsAt: string;
  createdAt: string;
}

export interface GiveawayInput {
  channelId: string;
  guildId: string;
  prize: string;
  winnersCount: number;
  durationMs: number;
  hostId: string;
  hostUsername: string;
}

export interface AuditLog {
  id: number;
  guildId: string;
  action: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  targetId?: string | null;
  /** @nullable */
  targetUsername?: string | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export type ListTicketsParams = {
guildId?: string;
status?: string;
};

export type ListWarningsParams = {
guildId?: string;
};

export type GetLeaderboardParams = {
guildId?: string;
};

export type ListGiveawaysParams = {
guildId?: string;
};

export type ListAuditLogsParams = {
guildId?: string;
};