# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  guildId: string;
  action: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  targetId?: string | null;
  /** @nullable */
  targetUsername?: string | null;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}