# lib/api-zod/src/generated/types/giveawayInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GiveawayInput {
  channelId: string;
  guildId: string;
  prize: string;
  winnersCount: number;
  durationMs: number;
  hostId: string;
  hostUsername: string;
}