# lib/api-zod/src/generated/types/guildSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildSettings {
  /** @nullable */
  id?: number | null;
  guildId: string;
  /** @nullable */
  welcomeEnabled?: boolean | null;
  /** @nullable */
  welcomeChannelId?: string | null;
  /** @nullable */
  welcomeTitle?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  /** @nullable */
  welcomeColor?: string | null;
  /** @nullable */
  welcomeShowAvatar?: boolean | null;
  /** @nullable */
  autoRoleEnabled?: boolean | null;
  /** @nullable */
  autoRoleId?: string | null;
  /** @nullable */
  ticketEnabled?: boolean | null;
  /** @nullable */
  ticketChannelId?: string | null;
  /** @nullable */
  ticketCategory?: string | null;
  /** @nullable */
  ticketMessage?: string | null;
  /** @nullable */
  loggingEnabled?: boolean | null;
  /** @nullable */
  loggingChannelId?: string | null;
  /** @nullable */
  moderationEnabled?: boolean | null;
  /** @nullable */
  moderationChannels?: string | null;
  /** @nullable */
  antiSpamEnabled?: boolean | null;
  /** @nullable */
  antiSpamThreshold?: string | null;
  /** @nullable */
  xpEnabled?: boolean | null;
  /** @nullable */
  xpPerMessage?: string | null;
  /** @nullable */
  bannedWords?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}