# lib/api-zod/src/generated/types/guildSettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildSettingsInput {
  welcomeEnabled?: boolean;
  /** @nullable */
  welcomeChannelId?: string | null;
  welcomeTitle?: string;
  welcomeMessage?: string;
  welcomeColor?: string;
  welcomeShowAvatar?: boolean;
  autoRoleEnabled?: boolean;
  /** @nullable */
  autoRoleId?: string | null;
  ticketEnabled?: boolean;
  /** @nullable */
  ticketChannelId?: string | null;
  ticketCategory?: string;
  ticketMessage?: string;
  loggingEnabled?: boolean;
  /** @nullable */
  loggingChannelId?: string | null;
  moderationEnabled?: boolean;
  moderationChannels?: string;
  antiSpamEnabled?: boolean;
  antiSpamThreshold?: string;
  xpEnabled?: boolean;
  xpPerMessage?: string;
  bannedWords?: string;
}