# lib/api-zod/src/generated/types/messageResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageResult {
  id: string;
  success: boolean;
}